// AuctionClient/src/types/User.ts
export interface User {
  id: string;
  name: string;
  surname: string;
  email: string;
  username: string;
  address: string;
  phone: string;
}
export interface UserCreation {
  name: string;
  surname: string;
  email: string;
  username: string;
  address: string;
  phone: string;
  password: string;
}